pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8.11"
azure-communication-sms = "^1.0.1"
certifi = "2020.4.5.1"
cffi = "1.14.0"
chardet = "3.0.4"
click = "8.0.0"
cryptography = "3.3.2"
docker = "4.2.0"
Flask = "1.1.2"
Flask-PyMongo = "2.3.0"
gunicorn = "20.0.4"
idna = "2.9"
itsdangerous = "1.1.0"
Jinja2 = "2.11.2"
MarkupSafe = "1.1.1"
pycparser = "2.20"
PyJWT = "1.7.1"
pymongo = "3.10.1"
pyOpenSSL = "19.1.0"
pyshp = "2.1.0"
python-http-client = "3.2.7"
pytz = "2020.1"
requests = "2.23.0"
sendgrid = "6.3.1"
six = "1.14.0"
urllib3 = "1.25.9"
Werkzeug = "1.0.1"
xkcdpass = "1.17.3"
python-dotenv = "^0.19.0"
oauthlib = "^3.1.1"
requests-oauthlib = "^1.3.0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-flask = "^1.2.0"
pytest-html = "^3.1.1"
pytest-metadata = "^1.11.0"
Flask-RESTful = "0.3.8"
pandas = "1.2.3"
Flask-Cors = "3.0.8"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
flake8-isort = "^4.0.0"
isort = "^5.7.0"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
addopts = "--doctest-modules"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
